returning fitting results now as list column in sDATA
restructured partGLFitLRCWindows to return fitting results in a
list-column instead of a list of fitting results and a simple
data.frame.

Fitting results of with and without VPD optimization are now stored in
resulting data.frame at the lines of parameter estimates.

To not confuse the user with the list columns, they are not exported by
default with sExportResult.