started moving to devtools
many problems with default exporting all objects that start with a . and with the order of sourcing files. Hence, wanted to make use of roxygen2 imports and export and includes to generate NAMESPACE file.

Non-Exported but tested objects need to be prefixed by package name in tests.

When adopting devtools, also changed workflow to loading library instead of sourcing R-files.

Some files and techniques became obsolete by this process.
This changeset involved creating a new package skeleton and subsequently copying the former contents to the new skeleton.

still TODO: man and examples and RCheck